Unify landing page readiness validation
Plan: plans/PR-Landing-Page-Readiness-Validator-Unification.md

Intentional:

- Keep the structural quality pack in place and add shared SEO/AEO/GEO readiness as an additional save-time gate.

- Preserve export and public robots readiness payload shapes.

Deferred:

- Saved-draft repair action, editable drafts, and public prerendering remain follow-up slices.
